requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
lxml==4.9.3
python-dotenv==1.0.0
typer==0.9.0
loguru==0.7.2
pytest==7.4.3
httpx==0.25.2
django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
chardet